% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barplot.R
\name{MIDEbar}
\alias{MIDEbar}
\title{Mide Barplots}
\usage{
MIDEbar(data, vars, varlabels, barcolor = MIDEpallete, shadowdata,
  horizontal = FALSE)
}
\arguments{
\item{data}{a dataframe with only one row}

\item{vars}{a character vector with the names of the columns to plot}

\item{varlabels}{character vector with the names to be used in the plot}

\item{barcolor}{list of colors to be used in the bars, defaults to MIDEpallete}

\item{shadowdata}{a second observation to plot, this one has a low alpha, recomended for highest case}

\item{horizontal}{Should the plot be horizontal?, defaults to FALSE}
}
\value{
a ggplot graph
}
\description{
This function takes one observation and makes varplots for
multiple variales, if allows comparison if a second observation is
passed as shadowdata
}
\details{
This function takes a single observation in a dataframe,
makes a barplot using as x values the variables gives in vars,
and as y values the values in the variables themselves. If a second
observation is given in shadowvar this one is shown with a low alpha
to be used as comparison.
}
\author{
Andrés Mejía
}
