% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barplot.R
\docType{data}
\name{MIDEpallete}
\alias{MIDEpallete}
\title{Colors used in the MIDE documentation}
\format{An object of class \code{character} of length 7.}
\usage{
MIDEpallete
}
\description{
List of Colors used in the MIDE documentation
}
\keyword{datasets}
