% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizontal plot.R
\name{rankingPlot}
\alias{rankingPlot}
\title{Ranking Plot}
\usage{
rankingPlot(data, name, num, limits, colores, n, paleta)
}
\arguments{
\item{data}{a dataframe with the info}

\item{name}{the name of a column to be used as the names of the elements}

\item{num}{the name of the numeric variable to be compared}

\item{limits}{range of the plot should be in the format c(min,max)}

\item{colores}{name of a column giving information of the color of each point}

\item{n}{show the top n cases}

\item{paleta}{colors to be used in the points}
}
\value{
a ggplot graph
}
\description{
This function makes a plot that is intented to be used for
comparison of the scores in a single variable
}
\details{
This function takes a dataframe and two columns, one of the name
and the other of a numeric attribute, this function orders the
data and makes a horizontal point plot that makes easy compare
the results between units.
}
\author{
Andrés Mejía
}
