% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Araña.R
\name{SpiderPlot}
\alias{SpiderPlot}
\title{Spider Plot}
\usage{
SpiderPlot(data, vars, varlabels, pos = 1:dim(data)[1])
}
\arguments{
\item{data}{a data frame}

\item{vars}{variables to use}

\item{pos}{index of the cases to plot, defaults to all}

\item{varlables}{labels to use in the plot in case those differ
from the colnames in data}
}
\value{
A list of ggplot graphs
}
\description{
Does a list of circular plots showing the position of the n-th
entry in a ranking, the
fist places being outside (a spider)
}
\details{
This function takes a dataframe and makes the spider plots
(a circular ranking plot) of the elements listed.
}
\author{
Andrés Mejía
}
